+++ vision/detectors/__init__.py
@@ -1,7 +1,9 @@
 from .color import ColorDetector
 from .face import FaceDetector
+from .object import ObjectDetector
 
 __all__ = [
     "ColorDetector",
     "FaceDetector",
+    "ObjectDetector",
 ]